Enum libalpm::Event

pub enum Event<'a> {
    CheckDepsStart,
    CheckDepsDone,
    FileConflictsStart,
    FileConflictsDone,
    ResolveDepsStart,
    ResolveDepsDone,
    InterConflictsStart,
    InterConflictsDone,
    TransactionStart,
    TransactionDone,
    PackageOperationStart(PackageOperation<'a>),
    PackageOperationDone(PackageOperation<'a>),
    IntegrityStart,
    IntegrityDone,
    LoadStart,
    LoadDone,
    DeltaIntegrityStart,
    DeltaIntegrityDone,
    DeltaPatchesStart,
    DeltaPatchesDone,
    DeltaPatchStart,
    DeltaPatchDone,
    DeltaPatchFailed,
    ScriptletInfo,
    RetrieveStart,
    RetrieveDone,
    RetrieveFailed,
    PkgDownloadStart,
    PkgDownloadDone,
    PkgDownloadFailed,
    DiskspaceStart,
    DiskspaceDone,
    OptDepRemoval,
    DatabaseMissing,
    KeyringStart,
    KeyringDone,
    KeyDownloadStart,
    KeyDownloadDone,
    PacnewCreated,
    PacsaveCreated,
    HookStart,
    HookDone,
    HookRunStart,
    HookRunDone,
    __Unknown,
}

An event emitted from libalpm

Variants
CheckDepsStart

Dependencies will be computed for a package.

CheckDepsDone

Dependencies were computed for a package.

FileConflictsStart

File conflicts will be computed for a package.

FileConflictsDone

File conflicts were computed for a package.

ResolveDepsStart

Dependencies will be resolved for target package.

ResolveDepsDone

Dependencies were resolved for target package.

InterConflictsStart

Inter-conflicts will be checked for target package.

InterConflictsDone

Inter-conflicts were checked for target package.

TransactionStart

Processing the package transaction is starting.

TransactionDone

Processing the package transaction is finished.

PackageOperationStart(PackageOperation<'a>)

Package will be installed/upgraded/downgraded/re-installed/removed; See alpm_event_package_operation_t for arguments.

PackageOperationDone(PackageOperation<'a>)

Package was installed/upgraded/downgraded/re-installed/removed; See alpm_event_package_operation_t for arguments.

IntegrityStart

Target package's integrity will be checked.

IntegrityDone

Target package's integrity was checked.

LoadStart

Target package will be loaded.

LoadDone

Target package is finished loading.

DeltaIntegrityStart

Target delta's integrity will be checked.

DeltaIntegrityDone

Target delta's integrity was checked.

DeltaPatchesStart

Deltas will be applied to packages.

DeltaPatchesDone

Deltas were applied to packages.

DeltaPatchStart

Delta patch will be applied to target package; See alpm_event_delta_patch_t for arguments..

DeltaPatchDone

Delta patch was applied to target package.

DeltaPatchFailed

Delta patch failed to apply to target package.

ScriptletInfo

Scriptlet has printed information; See alpm_event_scriptlet_info_t for arguments.

RetrieveStart

Files will be downloaded from a repository.

RetrieveDone

Files were downloaded from a repository.

RetrieveFailed

Not all files were successfully downloaded from a repository.

PkgDownloadStart

A file will be downloaded from a repository; See alpm_event_pkgdownload_t for arguments

PkgDownloadDone

A file was downloaded from a repository; See alpm_event_pkgdownload_t for arguments

PkgDownloadFailed

A file failed to be downloaded from a repository; See alpm_event_pkgdownload_t for arguments

DiskspaceStart

Disk space usage will be computed for a package.

DiskspaceDone

Disk space usage was computed for a package.

OptDepRemoval

An optdepend for another package is being removed; See alpm_event_optdep_removal_t for arguments.

DatabaseMissing

A configured repository database is missing; See alpm_event_database_missing_t for arguments.

KeyringStart

Checking keys used to create signatures are in keyring.

KeyringDone

Keyring checking is finished.

KeyDownloadStart

Downloading missing keys into keyring.

KeyDownloadDone

Key downloading is finished.

PacnewCreated

A .pacnew file was created; See alpm_event_pacnew_created_t for arguments.

PacsaveCreated

A .pacsave file was created; See alpm_event_pacsave_created_t for arguments

HookStart

Processing hooks will be started.

HookDone

Processing hooks is finished.

HookRunStart

A hook is starting

HookRunDone

A hook has finished running

__Unknown